backend/requirements.txt
flask==3.0.0
pyngrok==7.2.3
PyPDF2==3.0.1
google-generativeai==0.8.4